Login user credential insertion with MVP done

+++ app/src/main/java/net/dynu/w3rkaut/domain/model/User.java
@@ -5,23 +5,23 @@ package net.dynu.w3rkaut.domain.model;
  */
 
 public class User {
-    private int userId;
+    private long userId;
     private String email;
     private String firstName;
     private String lastName;
 
-    public User(int userId, String email, String firstName, String lastName) {
+    public User(long userId, String email, String firstName, String lastName) {
         this.userId = userId;
         this.email = email;
         this.firstName = firstName;
         this.lastName = lastName;
     }
 
-    public int getUserId() {
+    public long getUserId() {
         return userId;
     }
 
-    public void setUserId(int userId) {
+    public void setUserId(long userId) {
         this.userId = userId;
     }
